requirements.txt
pythonocc-core>=7.7.0
networkx>=3.0
matplotlib>=3.5.0
numpy>=1.21.0
scipy>=1.7.0
